# swagger-spec.json
{"openapi":"3.0.0","paths":{"/auth/registration":{"post":{"operationId":"AuthController_signUp","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignUpDto"}}}},"responses":{"201":{"description":"User created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignUpCreatedResponse"}}}},"400":{"description":"User validation errors","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignUpBadResponse"}}}}},"tags":["Authentication"]}},"/auth/login":{"post":{"operationId":"AuthController_signIn","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignInDto"}}}},"responses":{"200":{"description":"User successfully logged in","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignInOkResponse"}}}},"400":{"description":"Login or password is wrong","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignInBadResponse"}}}}},"tags":["Authentication"]}},"/auth/forgot/send-code":{"post":{"operationId":"AuthController_sendCode","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SendCodeDto"}}}},"responses":{"200":{"description":"Code sended to your email","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SendCodeOkResponse"}}}},"400":{"description":"User from this email not founded","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SendCodeBadResponse"}}}}},"tags":["Authentication"]}},"/auth/forgot/confirm-code":{"post":{"operationId":"AuthController_confirmCode","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConfirmCodeDto"}}}},"responses":{"200":{"description":"Right code","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConfirmCodeOkResponse"}}}},"400":{"description":"Wrong code","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConfirmCodeBadResponse"}}}}},"tags":["Authentication"]}},"/auth/forgot/change-password/{token}":{"put":{"operationId":"AuthController_changePassword","parameters":[{"name":"token","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChangePasswordDto"}}}},"responses":{"200":{"description":"Right code","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChangePasswordOkResponse"}}}},"400":{"description":"Wrong code","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChangePasswordBadResponse"}}}}},"tags":["Authentication"]}},"/user":{"get":{"operationId":"UserController_findMe","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"401":{"description":"Unauthorized"}},"tags":["User"],"security":[{"bearer":[]}]}},"/user/balance":{"get":{"operationId":"UserController_myBalance","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MyBalanceResponse"}}}},"401":{"description":"Unauthorized"}},"tags":["User"],"security":[{"bearer":[]}]}},"/user/bonus":{"post":{"operationId":"UserController_addBonus","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddBonusDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddBonusResponse"}}}},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddBonusBadResponse"}}}},"401":{"description":"Unauthorized"}},"tags":["User"],"security":[{"bearer":[]}]},"get":{"operationId":"UserController_getBonuses","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Bonus"}}}},"401":{"description":"Unauthorized"}},"tags":["User"],"security":[{"bearer":[]}]}},"/user/confirm-email/send-code":{"post":{"operationId":"UserController_confirmEmailSendCode","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConfirmEmailSendCodeResponse"}}}},"401":{"description":"Unauthorized"}},"tags":["User"],"security":[{"bearer":[]}]}},"/user/confirm-email":{"post":{"operationId":"UserController_confirmEmailConfirmCode","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConfirmCodeDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConfirmEmailConfirmCodeResponse"}}}},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConfirmEmailConfirmCodeBadResponse"}}}},"401":{"description":"Unauthorized"}},"tags":["User"],"security":[{"bearer":[]}]}},"/user/change-email/send-code":{"post":{"operationId":"UserController_confirm","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SendCodeDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConfirmEmailSendCodeResponse"}}}},"401":{"description":"Unauthorized"}},"tags":["User"],"security":[{"bearer":[]}]}},"/user/change-email":{"put":{"operationId":"UserController_confirmCode","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConfirmCodeDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChangeEmailConfirmCodeResponse"}}}},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConfirmEmailConfirmCodeBadResponse"}}}},"401":{"description":"Unauthorized"}},"tags":["User"],"security":[{"bearer":[]}]}},"/user/password/confirm":{"put":{"operationId":"UserController_oldPasswordConfirm","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OldPasswordConfirmDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignInOkResponse"}}}},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OldPasswordConfirmBadResponse"}}}},"401":{"description":"Unauthorized"}},"tags":["User"],"security":[{"bearer":[]}]}},"/user/password/{token}":{"put":{"operationId":"UserController_changePassword","parameters":[{"name":"token","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChangePasswordDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChangePasswordResponse"}}}},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChangePasswordBadResponse"}}}},"401":{"description":"Unauthorized"}},"tags":["User"],"security":[{"bearer":[]}]}},"/user/requisites":{"get":{"operationId":"UserController_findRequisites","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FindRequisitesResponse"}}}},"401":{"description":"Unauthorized"}},"tags":["User"],"security":[{"bearer":[]}]}},"/user/requisites/recommended":{"get":{"operationId":"UserController_findRecommendedRequisites","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Requisite"}}}},"401":{"description":"Unauthorized"}},"tags":["User"],"security":[{"bearer":[]}]}},"/bets/tops":{"get":{"operationId":"BetsController_topBets","parameters":[{"name":"skip","required":false,"in":"query","example":0,"description":"Skip elements","schema":{"default":0,"type":"number"}},{"name":"limit","required":false,"in":"query","example":6,"description":"Page limit","schema":{"default":6,"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Bet"}}}}},"tags":["Bets"]}},"/bets/my":{"get":{"operationId":"BetsController_myBets","parameters":[{"name":"skip","required":false,"in":"query","example":0,"description":"Skip elements","schema":{"default":0,"type":"number"}},{"name":"limit","required":false,"in":"query","example":6,"description":"Page limit","schema":{"default":6,"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Bet"}}}},"401":{"description":"Unauthorized"}},"tags":["Bets"],"security":[{"bearer":[]}]}},"/admin/login":{"post":{"operationId":"AdminController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdminLoginDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignInOkResponse"}}}}},"tags":["Admin"]}},"/admin/requisites":{"post":{"operationId":"AdminController_createRequisite","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateRequisiteDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Requisite"}}}},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateRequisiteBadResponse"}}}},"401":{"description":"Unauthorized"}},"tags":["Admin"],"security":[{"bearer":[]}]},"get":{"operationId":"AdminController_getRequisites","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Requisite"}}}},"401":{"description":"Unauthorized"}},"tags":["Admin"],"security":[{"bearer":[]}]}},"/admin/replenishments":{"get":{"operationId":"AdminController_getRequisite","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Replenishment"}}}},"401":{"description":"Unauthorized"}},"tags":["Admin"],"security":[{"bearer":[]}]}},"/admin/replenishments/{id}":{"put":{"operationId":"AdminController_confirmReplenishment","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConfirmReplenishmentResponse"}}}},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConfirmReplenishmentBadResponse"}}}},"401":{"description":"Unauthorized"}},"tags":["Admin"],"security":[{"bearer":[]}]}},"/replenishments":{"get":{"operationId":"ReplenishmentController_findAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Replenishment"}}}}},"401":{"description":"Unauthorized"}},"tags":["Replenishments"],"security":[{"bearer":[]}]},"post":{"operationId":"ReplenishmentController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateReplenishmentDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Replenishment"}}}},"401":{"description":"Unauthorized"}},"tags":["Replenishments"],"security":[{"bearer":[]}]}},"/replenishments/{id}":{"get":{"operationId":"ReplenishmentController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Replenishment"}}}},"401":{"description":"Unauthorized"}},"tags":["Replenishments"],"security":[{"bearer":[]}]}},"/replenishments/cancel/{id}":{"put":{"operationId":"ReplenishmentController_cancel","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CancelReplenishmentResponse"}}}},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CancelReplenishmentBadResponse"}}}},"401":{"description":"Unauthorized"}},"tags":["Replenishments"],"security":[{"bearer":[]}]}},"/replenishments/confirm/{id}":{"put":{"operationId":"ReplenishmentController_confirm","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConfirmReplenishmentResponse"}}}},"401":{"description":"Unauthorized"}},"tags":["Replenishments"],"security":[{"bearer":[]}]}}},"info":{"title":"Aviator","description":"Aviator API документация","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"SignUpDto":{"type":"object","properties":{"currency":{"type":"string","example":"RUB"},"login":{"type":"string","example":"test1234"},"password":{"type":"string","example":"12345678","minLength":8,"maxLength":30},"passwordConfirm":{"type":"string","example":"12345678"},"email":{"type":"string","example":"test@gmail.com"},"telegramId":{"type":"number","example":123456},"from":{"type":"string","example":"test123"}},"required":["currency","login","password","passwordConfirm","email","telegramId"]},"SignUpCreatedResponse":{"type":"object","properties":{"token":{"type":"string","example":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c","description":"Токен для авторизации"}},"required":["token"]},"SignUpBadResponse":{"type":"object","properties":{"status":{"type":"number","example":400},"message":{"example":["email is already used"],"type":"array","items":{"type":"string"}}},"required":["status","message"]},"SignInDto":{"type":"object","properties":{"login":{"type":"string","example":"test1234"},"password":{"type":"string","example":"123456"}},"required":["login","password"]},"SignInOkResponse":{"type":"object","properties":{"token":{"type":"string","example":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c","description":"Токен для авторизации"}},"required":["token"]},"SignInBadResponse":{"type":"object","properties":{"status":{"type":"number","example":400},"message":{"type":"string","example":"Логин или пароль неправильный"}},"required":["status","message"]},"SendCodeDto":{"type":"object","properties":{"email":{"type":"string","example":"test@gmail.com"}},"required":["email"]},"SendCodeOkResponse":{"type":"object","properties":{"message":{"type":"string","example":"Код отправлен на ваш email"}},"required":["message"]},"SendCodeBadResponse":{"type":"object","properties":{"status":{"type":"number","example":400},"message":{"type":"string","example":"Пользователь с таким email не найден"}},"required":["status","message"]},"ConfirmCodeDto":{"type":"object","properties":{"code":{"type":"string","example":"123456"}},"required":["code"]},"ConfirmCodeOkResponse":{"type":"object","properties":{"token":{"type":"string","example":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6IjY1NDI1ZDk2ZTg3ZmNkZTBmNTIzYWE1NCIsImlhdCI6MTY5ODkxMDQzNiwiZXhwIjoxNjk4OTk2ODM2fQ.3-CICCrb7jsNQjfgDqjJhnLB7K54sebRj44qdYEOzrg","description":"Токен для сброса пароля"}},"required":["token"]},"ConfirmCodeBadResponse":{"type":"object","properties":{"status":{"type":"number","example":400},"message":{"type":"string","example":"Неверный код"}},"required":["status","message"]},"ChangePasswordDto":{"type":"object","properties":{"password":{"type":"string","example":"12345678","minLength":8,"maxLength":30},"passwordConfirm":{"type":"string","example":"12345678"}},"required":["password","passwordConfirm"]},"ChangePasswordOkResponse":{"type":"object","properties":{"message":{"type":"string","example":""}},"required":["message"]},"ChangePasswordBadResponse":{"type":"object","properties":{"message":{"example":["Поле обязательно для заполнения"],"type":"array","items":{"type":"string"}},"error":{"type":"string","example":"Bad Request"},"statusCode":{"type":"number","example":400}},"required":["message","error","statusCode"]},"User":{"type":"object","properties":{"_id":{"type":"string"},"telegramId":{"type":"number"},"currency":{"type":"string"},"login":{"type":"string"},"email":{"type":"string"},"balance":{"type":"number"},"referralBalance":{"type":"number"},"leader":{"$ref":"#/components/schemas/User"},"descendants":{"type":"array","items":{"type":"string"}},"bonuses":{"type":"array","items":{"type":"string"}}},"required":["_id","telegramId","currency","login","email","balance","referralBalance","leader","descendants","bonuses"]},"MyBalanceResponse":{"type":"object","properties":{"balance":{"type":"number"},"currency":{"type":"string"}},"required":["balance","currency"]},"AddBonusDto":{"type":"object","properties":{"promoCode":{"type":"string","example":"NewYear2099"}},"required":["promoCode"]},"AddBonusResponse":{"type":"object","properties":{"message":{"type":"string","example":"Промокод успешно активирован"}},"required":["message"]},"AddBonusBadResponse":{"type":"object","properties":{"message":{"type":"string","examples":["Нет такого промокода","такой промокод у вас уже есть"]}},"required":["message"]},"Bonus":{"type":"object","properties":{"_id":{"type":"string"},"type":{"type":"number"},"promoCode":{"type":"string"},"currency":{"type":"string"},"bonus":{"type":"number"},"bonusPercent":{"type":"number"},"bonusCoeff":{"type":"number"},"maxUsedCount":{"type":"number"},"usedCount":{"type":"number"},"expiresIn":{"format":"date-time","type":"string"},"users":{"type":"array","items":{"type":"string"}}},"required":["_id","type","promoCode","currency","bonus","bonusPercent","bonusCoeff","maxUsedCount","usedCount","expiresIn"]},"ConfirmEmailSendCodeResponse":{"type":"object","properties":{"message":{"type":"string","example":"На ваш Email отправлен код для подтверждения"}},"required":["message"]},"ConfirmEmailConfirmCodeResponse":{"type":"object","properties":{"message":{"type":"string","example":"Ваш email подтверждён"}},"required":["message"]},"ConfirmEmailConfirmCodeBadResponse":{"type":"object","properties":{"message":{"type":"string","example":"Неверный код"},"error":{"type":"string","example":"Bad Request"},"statusCode":{"type":"number","example":400}},"required":["message","error","statusCode"]},"ChangeEmailConfirmCodeResponse":{"type":"object","properties":{"message":{"type":"string","example":"Ваш email успешно изменен"}},"required":["message"]},"OldPasswordConfirmDto":{"type":"object","properties":{"password":{"type":"string","example":"12345678"}},"required":["password"]},"OldPasswordConfirmBadResponse":{"type":"object","properties":{"message":{"type":"string","example":"Неверный пароль"},"error":{"type":"string","example":"Bad Request"},"statusCode":{"type":"number","example":400}},"required":["message","error","statusCode"]},"ChangePasswordResponse":{"type":"object","properties":{"message":{"type":"string","example":"Ваш пароль успешно изменен"}},"required":["message"]},"Requisite":{"type":"object","properties":{"_id":{"type":"string"},"requisite":{"type":"string"},"name":{"type":"string"},"currency":{"type":"string"},"img":{"type":"string"},"commission":{"type":"number"},"status":{"type":"string","enum":["Активный","Неактивный"]}},"required":["_id","requisite","name","currency","img","commission","status"]},"FindRequisitesResponse":{"type":"object","properties":{"requisites":{"type":"array","items":{"$ref":"#/components/schemas/Requisite"}},"currency":{"type":"string","example":"USD"}},"required":["requisites","currency"]},"Bet":{"type":"object","properties":{"bet":{"type":"number"},"currency":{"type":"string","example":"USD"},"time":{"format":"date-time","type":"string"},"coeff":{"type":"number"},"win":{"type":"number"},"player":{"type":"string"}},"required":["bet","currency","time","coeff","player"]},"AdminLoginDto":{"type":"object","properties":{"login":{"type":"string","example":"test123"},"password":{"type":"string","example":"123456"}},"required":["login","password"]},"CreateRequisiteDto":{"type":"object","properties":{"requisite":{"type":"string","example":"0000-0000-0000-0000"}},"required":["requisite"]},"CreateRequisiteBadResponse":{"type":"object","properties":{"message":{"type":"string","example":"Такой реквизит уже существует"},"error":{"type":"string","example":"Bad Request"},"status":{"type":"number","example":400}},"required":["message","error","status"]},"Replenishment":{"type":"object","properties":{"user":{"type":"string"},"amount":{"type":"number","example":100},"currency":{"type":"string","example":"USD"},"deduction":{"type":"number","example":110},"status":{"type":"string","enum":["Ожидает оплаты","Успешно завершена","Отменена"]},"statusMessage":{"type":"string"},"isPayConfirmed":{"type":"boolean"},"requisite":{"oneOf":[{"type":"string"},{"$ref":"#/components/schemas/Requisite"}],"allOf":[{"$ref":"#/components/schemas/Requisite"}]},"createdAt":{"format":"date-time","type":"string"},"completedDate":{"format":"date-time","type":"string"}},"required":["user","amount","currency","deduction","status","statusMessage","isPayConfirmed","requisite","createdAt","completedDate"]},"ConfirmReplenishmentResponse":{"type":"object","properties":{"message":{"type":"string","example":"Оплата подтверждена"}},"required":["message"]},"ConfirmReplenishmentBadResponse":{"type":"object","properties":{"message":{"type":"string","example":"Вы не можете менять подтвержденную заявку"},"error":{"type":"string","example":"Bad Request"},"status":{"type":"number","example":400}},"required":["message","error","status"]},"CreateReplenishmentDto":{"type":"object","properties":{"currency":{"type":"string","example":"UZS"},"amount":{"type":"number","example":10000},"requisite":{"type":"string"}},"required":["currency","amount","requisite"]},"CancelReplenishmentResponse":{"type":"object","properties":{"message":{"type":"string","example":"Пополнение отменена"}},"required":["message"]},"CancelReplenishmentBadResponse":{"type":"object","properties":{"message":{"type":"string","example":"Невозможно отменить"},"error":{"type":"string","example":"Bad Request"},"status":{"type":"number","example":400}},"required":["message","error","status"]}}}}